+++ app/src/main/AndroidManifest.xml
@@ -24,13 +24,13 @@
             android:name=".RegisterCareActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="false" />
         <activity
             android:name="org.retriever.dailypet.ui.login.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".ui.splash.SplashActivity"
             android:exported="true"
             android:theme="@style/Splash">
             <intent-filter>
