+++ app/src/main/java/org/retriever/dailypet/ui/login/CreateProfileFragment.kt
@@ -31,6 +31,7 @@ import org.retriever.dailypet.R
 import org.retriever.dailypet.databinding.FragmentCreateProfileBinding
 import org.retriever.dailypet.model.Resource
 import org.retriever.dailypet.ui.base.BaseFragment
+import org.retriever.dailypet.ui.bottomsheet.FragmentCameraSheet
 import org.retriever.dailypet.util.hideProgressCircular
 import org.retriever.dailypet.util.showProgressCircular
