+++ app/src/main/java/org/retriever/dailypet/ui/main/MainActivity.kt
@@ -1,10 +1,11 @@
-package org.retriever.dailypet
+package org.retriever.dailypet.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
 import dagger.hilt.android.AndroidEntryPoint
+import org.retriever.dailypet.R
 import org.retriever.dailypet.databinding.ActivityMainBinding
 import org.retriever.dailypet.fragments.*
