+++ app/src/main/res/navigation/login_navigation_graph.xml
@@ -48,7 +48,7 @@
     </fragment>
     <activity
         android:id="@+id/mainActivity"
-        android:name="org.retriever.dailypet.MainActivity"
+        android:name="org.retriever.dailypet.ui.main.MainActivity"
         android:label="activity_main"
         tools:layout="@layout/activity_main" />
     <fragment
