+++ app/src/main/res/values/strings.xml
@@ -23,4 +23,5 @@
     <string name="invalid_petname_text">올바른 반려동물 이름을 입력해주세요.</string>
     <string name="valid_petname_text">사용가능한 반려동물 이름입니다.</string>
     <string name="already_used_petname_text">그룹에 이미 존재하는 반려동물 이름입니다.</string>
+    <string name="input_pet_birth_text">반려동물의 생일을 선택해주세요.</string>
 </resources>
